+++ src/FSharp.Control.AsyncSeq.Profile7/AssemblyInfo.fs
@@ -4,10 +4,10 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("FSharp.Control.AsyncSeq.Profile7")>]
 [<assembly: AssemblyProductAttribute("FSharp.Control.AsyncSeq")>]
 [<assembly: AssemblyDescriptionAttribute("Asynchronous sequences for F#")>]
-[<assembly: AssemblyVersionAttribute("2.0.11")>]
-[<assembly: AssemblyFileVersionAttribute("2.0.11")>]
+[<assembly: AssemblyVersionAttribute("2.0.13")>]
+[<assembly: AssemblyFileVersionAttribute("2.0.13")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "2.0.11"
-    let [<Literal>] InformationalVersion = "2.0.11"
+    let [<Literal>] Version = "2.0.13"
+    let [<Literal>] InformationalVersion = "2.0.13"